sửa controller restaurant

+++ backend/src/restaurant/restaurant.service.ts
@@ -12,23 +12,31 @@ export class RestaurantService {
        
     }
     create(createRestaurantDto: CreateRestaurantDto) {
-        // console.log(createRestaurantDto.)
         return this.Restaurantrepository.save(createRestaurantDto)
     }
 
-    findAll() {
-        return this.Restaurantrepository.find();
+    findMyRestaurant(idU: number) {
+        return this.Restaurantrepository.createQueryBuilder('restaurant').where('idUser = :idU', {idU: idU}).getMany();
     }
 
-    findOne(id:number): Promise<Restaurant> {
-        return this.Restaurantrepository.findOneBy({id})
+
+    async update(idU: number, id:number, updateRestaurantDto: {name: string, address: string, phoneNumber: string, latitude: number, longitude: number}) {
+        const isOwner = await this.Restaurantrepository.createQueryBuilder('restaurant').where('idUser = :idU', {idU: idU}).andWhere('id = :id', {id: id}).getOne();
+        if(isOwner) {
+            return this.Restaurantrepository.update(id, updateRestaurantDto)
+        } else {
+            return "Bạn không có quyền chỉnh sửa thông tin nhà hàng này";
+        }
+        
     }
 
-    update(id:number, updateRestaurantDto: UpdateRestaurantDto) {
-        return this.Restaurantrepository.update(id, updateRestaurantDto);
-    }
-
-    remove(id: number) {
-        this.Restaurantrepository.delete(id);
+    async remove(idU: number, id: number) {
+        const isOwner = await this.Restaurantrepository.createQueryBuilder('restaurant').where('idUser = :idU', {idU: idU}).andWhere('id = :id', {id: id}).getOne();
+        if(isOwner) {
+            this.Restaurantrepository.delete(id);
+            return "Xóa nhà hàng thành công";
+        } else {
+            return "Bạn không có quyền xóa thông tin nhà hàng này";
+        }
     }
 }
